requirements.txt
Django
Pillow
djangorestframework
python-decouple
drf-yasg
bs4
requests    
